moving around samples
