+++ sim/sample.py
@@ -1,15 +1,6 @@
 from utils import run
-import errno
-import math 
 import os
-import glob
 
-from simulated_genome import SimulatedGenome
-
-DEFAULT_REFERENCE_PATH = './Mycobacterium_bovis_AF212297_LT78304.fa'
-DEFAULT_MASK_PATH = './Mycbovis-2122-97_LT708304.fas.rpt.regions'
-DEFAULT_NUM_READ_PAIRS = 289994
-DEFAULT_NUM_READ_PAIRS = 100
 
 class Sample:
     @property
@@ -22,7 +13,8 @@ class Sample:
     def _simulate_genome_base(self, reference_path, simulated_genome_path, params):
         cmd = ["simuG.pl",
             "-refseq", reference_path,
-            "-prefix", simulated_genome_path + ".simulated"]
+            "-prefix", simulated_genome_path + ".simulated"
+        ]
         cmd.extend(params)
         run(cmd)
     
@@ -91,188 +83,3 @@ class Sample:
         os.rename(dwgsim_read_2, read_2)
 
         return read_1, read_2
-
-
-class VcfSample(Sample):
-    def __init__(self, 
-        predef_snp_path, 
-        reference_path=DEFAULT_REFERENCE_PATH,
-        seed=1, 
-        per_base_error_rate="0",
-        num_read_pairs = DEFAULT_NUM_READ_PAIRS,
-    ):
-
-        # TODO: Validate
-        self.reference_path = reference_path
-        
-        self.seed = seed
-        self.per_base_error_rate = per_base_error_rate
-        self.num_read_pairs = math.ceil(num_read_pairs)
-        
-        if os.path.isfile(predef_snp_path):
-            self.predef_snp_path = predef_snp_path
-        else:
-            raise Exception("Cant find the pre-defined snp path")
-        
-    @property
-    def name(self):
-        # extract filename
-        basename = os.path.splitext(os.path.basename(self.predef_snp_path))[0]
-        #TODO: this pattern is getting somewhat unweildy.
-        #   etiher use something more general or save metadata separately
-        return f"{type(self).__name__}-{basename}-seed{self.seed}-error{self.per_base_error_rate}-readpairs-{self.num_read_pairs}"
-
-    def decompose_complex_snps(self, output_file_path):
-        run(['vt', 
-            'decompose_blocksub',
-            self.predef_snp_path,
-            '-o', output_file_path])
-
-    def simulate_genome(self, simulated_genome_path, seed=1):
-        """ Simulated a genome with random SNPs
-
-            TODO: rename simulated_genome_path to simulated_genome_prefix
-
-            Parameters:
-                reference_path (str): Path to reference genome
-                simulated_genome_path (str): Path to simlated genome
-                predef_snps_path (str): Path to snippy generated VCF file.
-                seed (int): Seed value for simulation
-
-            Returns:
-                None
-        """
-        if not os.path.isfile(self.predef_snp_path):
-            raise(FileNotFoundError(errno.ENOENT, os.strerror(errno.ENOENT), self.predef_snp_path))
-        tmp_decomposed_vcf_path = simulated_genome_path + '.decomposed.vcf'
-        # Decompose complex SNPs
-        self.decompose_complex_snps(tmp_decomposed_vcf_path)
-        params = ["-snp_vcf", tmp_decomposed_vcf_path, 
-                  "-indel_vcf", tmp_decomposed_vcf_path,
-                  "-seed", str(seed)]
-        self._simulate_genome_base(self.reference_path, simulated_genome_path, params)
-        # clean tmp decomposed VCF
-        run(['rm', tmp_decomposed_vcf_path])
-
-        # TODO: DRY with RandomSample
-        snp_vcf_path = simulated_genome_path + '.simulated.refseq2simseq.SNP.vcf'
-        indel_vcf_path = simulated_genome_path + '.simulated.refseq2simseq.INDEL.vcf'
-        snp_table_path = simulated_genome_path + '.simulated.refseq2simseq.map.txt'
-        genome_path = simulated_genome_path + '.simulated.simseq.genome.fa'
-
-        return SimulatedGenome(self.name, genome_path, snp_table_path, snp_vcf_path, indel_vcf_path)
-
-
-class RandomSample(Sample):
-    def __init__(self, 
-        reference_path=DEFAULT_REFERENCE_PATH,
-        num_snps=16000, 
-        num_indels=3898, 
-        seed=1, 
-        per_base_error_rate="0",
-        num_read_pairs = DEFAULT_NUM_READ_PAIRS
-    ):
-        # TODO: Validate
-        self.reference_path = reference_path
-
-        self.num_snps = num_snps
-        self.num_indels = num_indels
-        self.seed = seed
-        self.per_base_error_rate = per_base_error_rate
-        self.num_read_pairs = math.ceil(num_read_pairs)
-        
-    @property
-    def name(self):
-        return f"{type(self).__name__}-snps{self.num_snps}-indels{self.num_indels}-seed{self.seed}"
-
-    def simulate_genome(self, simulated_genome_path):
-        """ Simulated a genome with random SNPs
-
-            TODO: rename simulated_genome_path to simulated_genome_prefix
-            Parameters:
-                reference_path (str): Path to reference genome
-                simulated_genome_path (str): Path to simlated genome
-                num_snps (int): Number of random SNPs
-                seed (int): Seed value for simulation
-
-            Returns:
-                None
-        """
-        params = [
-            "-snp_count", str(self.num_snps),
-            "-indel_count", str(self.num_indels),
-            "-seed", str(self.seed)
-        ]
-
-        self._simulate_genome_base(self.reference_path, simulated_genome_path, params)
-
-        # TODO: DRY with VcfSample
-        snp_vcf_path = simulated_genome_path + '.simulated.refseq2simseq.SNP.vcf'
-        indel_vcf_path = simulated_genome_path + '.simulated.refseq2simseq.INDEL.vcf'
-        snp_table_path = simulated_genome_path + '.simulated.refseq2simseq.map.txt'
-        genome_path = simulated_genome_path + '.simulated.simseq.genome.fa'
-
-        return SimulatedGenome(self.name, genome_path, snp_table_path, snp_vcf_path, indel_vcf_path)
-
-def vcf_samples(datasets, snippy_dir='/mnt/fsx-027/snippy/'):
-    """ Returns a list of VcfSample objects built from data sets contained
-        in the datasets parameter. Each object is instantiated with a 
-        unique seed value
-
-        Parameters:
-            datasets (tuple (str)): elements describe the base directory name
-                for each dataset
-            snippy_dir (str): path of parent directory of datasets
-    """
-    vcf_filepaths = []
-    for set_i in datasets:
-        vcf_dir = snippy_dir + set_i
-        if not os.path.isdir(vcf_dir):
-            raise Exception("Predefined SNP directory not found") 
-        vcf_dir = os.path.join(vcf_dir, '')
-        vcf_filepaths.extend(glob.glob(vcf_dir+'*.vcf'))
-    
-    # sort vcf_filepaths for using consistent seed values accross runs
-    vcf_filepaths.sort()
-    samples = []
-    seed_value = 0
-    for filepath in vcf_filepaths:
-        seed_value+=1 # different seed value for each sample
-        samples.append(VcfSample(filepath, seed=seed_value, per_base_error_rate="0.001-0.01"))    
-
-    return samples
-
-def quick_samples():
-    """ Returns a list containing a single RandomSample object
-    """
-    return [RandomSample(seed=1)]
-
-def random_samples():
-    """ Returns a list containing two RandomSample objects 
-    """
-    return [RandomSample(seed=5, num_snps=0, num_indels=0, per_base_error_rate="0.001-0.01"),
-            RandomSample(seed=1, per_base_error_rate="0.001-0.01"),
-            RandomSample(seed=666, per_base_error_rate="0.001-0.01")]
-
-def standard_samples():
-    """ Returns a list of Sample objects for the standard dataset
-    """
-    samples = vcf_samples(('standard',)) + random_samples() 
-    return samples
-
-def aph_samples():
-    """ Returns a list of Sample objects for the aph dataset
-    """
-    return vcf_samples(('aph',))
-
-def zwyer_samples():
-    """ Returns a list of Sample objects for the zwyer dataset 
-    """
-    return vcf_samples(('zwyer',))
-
-def all_samples():
-    """ Returns a list of Sample objects for all VCF samples and two
-        random samples
-    """
-    samples = vcf_samples(('aph', 'zwyer')) + random_samples()
-    return samples
