+++ sim/sequenced.py
