--- sim/simulated_sample.py
@@ -1,4 +1,0 @@
-class SimulatedSample:
-    def __init__(self):
-        self.
-
